Simplify MathJax bundle structure
