+++ bokehjs/src/lib/external/mathjax.d.ts
@@ -1,5 +1,5 @@
 declare namespace MathJax {
   function tex2svg(input: string): HTMLElement
-  // function ascii2svg(input: string): HTMLElement
+  function ascii2svg(input: string): HTMLElement
   function mathml2svg(input: string): HTMLElement
 }
